pyproject.toml
[tool.poetry]
name = "mood2mix"
version = "0.1.0"
description = "Emotion-based playlist generator using FastAPI and OpenAI"
authors = ["Vadym <67538058+Extractoro@users.noreply.github.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = ">=0.115.0,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
openai = ">=1.8.0,<2.0.0"
python-dotenv = ">=1.0.0,<2.0.0"
pydantic = "^2.11.7"
requests = "^2.32.4"
ytmusicapi = "^1.10.3"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
